Allow setting seed in Properties main class
