+++ sbt-test/shared/src/main/scala/hedgehog/sbt/Framework.scala
@@ -7,7 +7,6 @@ import hedgehog.core._
 import hedgehog.runner._
 import _root_.sbt.{testing => sbtt}
 import org.portablescala.reflect.Reflect
-import scala.util.Try
 
 class Framework extends sbtt.Framework {
 
@@ -93,7 +92,7 @@ class Task(
 
   private def run(eventHandler: sbtt.EventHandler, loggers: Array[sbtt.Logger]) = {
     val config = PropertyConfig.default
-    val seed = getSeed(loggers)
+    val seed = Seed.fromEnvOrTime(loggers.map(logger => logger.info(_)))
     val properties =
       if (fingerprint.isModule) {
         Reflect.lookupLoadableModuleClass(taskDef.fullyQualifiedName + "$", testClassLoader).get.loadModule().asInstanceOf[Properties]
@@ -108,20 +107,6 @@ class Task(
 
       loggers.foreach(logger => logger.info(Test.renderReport(taskDef.fullyQualifiedName, t, report, logger.ansiCodesSupported)))
     })
-  }
-
-  private def getSeed(loggers: Array[sbtt.Logger]): Seed = {
-    val (seed, logOutput) = sys.env
-      .get("HEDGEHOG_SEED")
-      .flatMap(s => Try(s.toLong).toOption) match {
-        case Some(seedFromEnv) => 
-          (seedFromEnv, s"Using seed from environment variable HEDGEHOG_SEED: $seedFromEnv")
-        case None =>
-          val seedFromTime = System.nanoTime()
-          (seedFromTime, s"Using random seed: $seedFromTime")
-      }
-    loggers.foreach(logger => logger.info(logOutput))
-    Seed.fromLong(seed)
   }
 }
